entity 생성 완료

+++ src/main/java/com/green/muziuniv_be_notuser/entity/semester/Semester.java
@@ -22,7 +22,7 @@ import java.time.Year;
 public class Semester {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long semesterId;
+    private int semesterId;
 
     @Column(nullable = false, columnDefinition = "YEAR")
     private Year year;
